+++ src/index.tsx
@@ -25,6 +25,8 @@ export * from './components/UnorderedList';
 
 export * from './components/Input';
 
+export * from './components/Select';
+
 export * from './components/Button';
 
 // Hooks